Respect !!str tag for structpb unmarshal (#50)
I observed that when a `structpb.Value_StringValue` can be parsed as a
number it does not roundtrip through protoyaml.

This commit adds some failing test cases to `TestValue_Dynamic` and
updates `unmarshalScalarString` to return a `Value_StringValue` if the
`!!str` tag is present.

Let me know if I'm missing something, I'm new to working with YAML!